+++ app/models/inventory.rb
@@ -3,4 +3,8 @@ class Inventory < ApplicationRecord
   belongs_to :survivor
 
   accepts_nested_attributes_for :stacks
+
+  def value
+    stacks.map(&:value).inject(0, :+)
+  end
 end